Update garethGallery.js
More changes, got the timeout working. Much more work needed on the
thumbs. Need to prevent double click on thumb navigation (as we can
never be sure of pane width. also link in touch events and auto pan
when on item that may be half over also need it to auto pan 
when move on gallery that thumbs control ... lots then!