Update garethGallery.js
Added progress bar support, and made it so user can define width, height
of responsive image swapping. Also can change the auto move speed and 
un-setting of the main timer is now done in a new function. Lastly
the width adjustments are made to the children as oppose to the images
this was bad as shown in the more complex gallery structure demo on
the demo's page GF